Add files via upload
You are going to need to download the most recent chrome driver and then add your own path @ the self.driver_path line. Here is the link to download the driver: https://googlechromelabs.github.io/chrome-for-testing/ . Make sure the binary is chromedriver and it is the most recent one. Also, the code does not work in Colab with file paths so I ran mine in jupyter notebook, but it was causing my computer to crash. Code works but haven't gotten it all to run yet. Let me know if it's better for you guys.